+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed policy check
 - Added `list-policy` command
+- Added `reset-object-acl` command
 - Added support for wildcards
 - Added support for customer-provided encryption key
 - Added `--version` option